feat(TrainAgentPerception): define perception concept and setup train perceptions
