+++ src/main/scala/ulisse/adapters/input/SimulationInfoAdapter.scala
@@ -4,7 +4,7 @@ import ulisse.applications.ports.SimulationInfoPorts
 import ulisse.entities.route.RouteEnvironmentElement
 import ulisse.entities.route.Routes.Route
 import ulisse.entities.station.{Station, StationEnvironmentElement}
-import ulisse.entities.train.TrainAgent
+import ulisse.entities.train.TrainAgents.TrainAgent
 import ulisse.entities.train.Trains.Train
 
 import scala.concurrent.Future
