+++ src/main/scala/ulisse/applications/ports/SimulationInfoPorts.scala
@@ -3,7 +3,7 @@ package ulisse.applications.ports
 import ulisse.entities.route.RouteEnvironmentElement
 import ulisse.entities.route.Routes.Route
 import ulisse.entities.station.{Station, StationEnvironmentElement}
-import ulisse.entities.train.TrainAgent
+import ulisse.entities.train.TrainAgents.TrainAgent
 import ulisse.entities.train.Trains.Train
 
 import scala.concurrent.Future
