+++ src/test/scala/ulisse/adapters/input/SimulationInfoAdapterTest.scala
@@ -9,7 +9,7 @@ import ulisse.applications.useCases.SimulationInfoService
 import ulisse.entities.route.RouteEnvironmentElement
 import ulisse.entities.route.Routes.Route
 import ulisse.entities.station.{Station, StationEnvironmentElement}
-import ulisse.entities.train.TrainAgent
+import ulisse.entities.train.TrainAgents.TrainAgent
 import ulisse.entities.train.Trains.Train
 
 import scala.concurrent.{Await, Future}
