+++ src/test/scala/ulisse/entities/station/PlatformTest.scala
@@ -5,7 +5,7 @@ import org.scalatest.matchers.should.Matchers
 import org.scalatest.wordspec.AnyWordSpec
 import ulisse.entities.simulation.EnvironmentElements.TrainAgentsContainer
 import ulisse.entities.simulation.EnvironmentElements.TrainAgentsDirection.Forward
-import ulisse.entities.train.TrainAgent
+import ulisse.entities.train.TrainAgents.TrainAgent
 import ulisse.entities.train.Trains.{Train, TrainTechnology}
 import ulisse.entities.train.Wagons.{UseType, Wagon}
