+++ src/test/scala/ulisse/entities/train/TrainAgentTest.scala
@@ -2,6 +2,7 @@ package ulisse.entities.train
 
 import org.scalatest.matchers.should.Matchers
 import org.scalatest.wordspec.AnyWordSpec
+import ulisse.entities.train.TrainAgents.TrainAgent
 import ulisse.entities.train.Trains.{Train, TrainTechnology}
 import ulisse.entities.train.Wagons.{UseType, Wagon}
